+++ plugins/postgresql/versions/16/install.sh
@@ -13,7 +13,7 @@ sysName=`uname`
 install_tmp=${rootPath}/tmp/mw_install.pl
 postgreDir=${serverPath}/source/postgresql
 
-VERSION=16.3
+VERSION=16.4
 
 # su - postgres -c "/www/server/postgresql/bin/pg_ctl start -D /www/server/postgresql/data"
 
@@ -96,8 +96,29 @@ Install_App()
 
 Uninstall_App()
 {
-	rm -rf $serverPath/postgresql
-	echo '卸载完成'
+	if [ -f /usr/lib/systemd/system/postgresql.service ];then
+		systemctl stop postgresql
+		systemctl disable postgresql
+		rm -rf /usr/lib/systemd/system/postgresql.service
+		systemctl daemon-reload
+	fi
+
+	if [ -f /lib/systemd/system/postgresql.service ];then
+		systemctl stop postgresql
+		systemctl disable postgresql
+		rm -rf /lib/systemd/system/postgresql.service
+		systemctl daemon-reload
+	fi
+
+	if [ -f $serverPath/postgresql/initd/postgresql ];then
+		$serverPath/postgresql/initd/postgresql stop
+	fi
+
+	if [ -d $serverPath/postgresql ];then
+		rm -rf $serverPath/postgresql
+	fi
+
+	echo '卸载[postgresql]完成'
 }
 
 action=$1
